pyproject.toml
[project]
name = "lab1"
version = "0.1.0"
description = ""
authors = [
    {name = "domster704",email = "domster704@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp (>=3.12.15,<4.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "fastapi[standart] (>=0.116.2,<0.117.0)"
]

[tool.poetry]
packages = [
    { include = "src" },
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)"
]
